restructuring project's folder and refactoring code into controller and route folder
