+++ routes/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  qr: require("./qr.routes"),
+};
