+++ routes/qr.routes.js
@@ -1,0 +1,6 @@
+const QrController = require("../controllers/QrController");
+
+module.exports = function(router) {
+  router.get("/", QrController.index);
+  router.get("/render", QrController.renderQR);
+};
